fix: double invalidate bug
Currently, if a query gets invalidated while its state is Loading(Some(value)), it loses the value. This can be annoying in cases where the query is invalidated multiple times within a short timespan, or for long-running queries.

We can fix this by nooping set_loading calls for queries that are already loading, which is what this change adds.